Refactor: 마일리지 카테고리 게시판 리펙터링 완료 (#4)
