+++ src/assets/data/board/board.ts
@@ -1,4 +1,3 @@
-import { MAX_MILEAGE, MANAGE, CHECK_BOX, NUM, CATEGORY } from './fields';
 import Checkbox from '@mui/material/Checkbox';
 /**
  * @breif 게시판 종류 정의
@@ -34,15 +33,3 @@ export const BoardList = [
   '마일리지 선정결과',
   '설정',
 ];
-
-/**
- * @breif [마일리지 카테고리] 게시판
- */
-
-export enum MileageCategoryBoard {
-  'NUM' = NUM,
-  'CATEGORY' = CATEGORY,
-  'MAX_MILEAGE' = MAX_MILEAGE,
-  'MANAGE' = MANAGE,
-  'CHECK_BOX' = CHECK_BOX,
-}
