+++ src/components/common/CustomTable.tsx
@@ -24,7 +24,7 @@ import { visuallyHidden } from '@mui/utils';
 import { StarIcon } from 'src/theme/overrides/CustomIcons';
 import { TablePagination } from '@mui/material';
 import CustomTablePagination from './Table/CustomTablePagination';
-import { MileageCategoryBoard } from '../../assets/data/board';
+import { MileageCategoryBoard } from '../../assets/data/board/mileageCategoryBoard';
 import { CATEGORY, NUM } from '../../assets/data/fields';
 
 // const rows = [
@@ -134,84 +134,84 @@ interface HeadCell {
  * @kind 마일리지 카테로리
  * @breif 데이터 인터페이스
  */
-interface Data {
-  [MileageCategoryBoard.CATEGORY]: string;
-  [MileageCategoryBoard.MAX_MILEAGE]: number;
-  [MileageCategoryBoard.MANAGE]: string;
-  [MileageCategoryBoard.CHECK_BOX]: string;
-}
+// interface Data {
+//   [MileageCategoryBoard.CATEGORY]: string;
+//   [MileageCategoryBoard.MAX_MILEAGE]: number;
+//   [MileageCategoryBoard.MANAGE]: string;
+//   [MileageCategoryBoard.CHECK_BOX]: string;
+// }
 
 /**
  * @kind 마일리지 카테고리
  * @brief 데이터 생성 함수
  *
  *  */
-function createData(
-  num: number,
-  category: string,
-  maxMileage: number,
-  manage: string,
-  checkBox: string
-): Data {
-  return {
-    [MileageCategoryBoard.NUM]: num,
-    [MileageCategoryBoard.CATEGORY]: category,
-    [MileageCategoryBoard.MAX_MILEAGE]: maxMileage,
-    [MileageCategoryBoard.MANAGE]: manage,
-    [MileageCategoryBoard.CHECK_BOX]: checkBox,
-  };
-}
+// function createData(
+//   num: number,
+//   category: string,
+//   maxMileage: number,
+//   manage: string,
+//   checkBox: string
+// ): Data {
+//   return {
+//     [MileageCategoryBoard.NUM]: num,
+//     [MileageCategoryBoard.CATEGORY]: category,
+//     [MileageCategoryBoard.MAX_MILEAGE]: maxMileage,
+//     [MileageCategoryBoard.MANAGE]: manage,
+//     [MileageCategoryBoard.CHECK_BOX]: checkBox,
+//   };
+// }
 
 /**
  * @number 1번 헤더
  * @description 마일리지 카테고리 리스트
  */
-const headCells = [
-  {
-    id: [MileageCategoryBoard.NUM],
-    numeric: false,
-    disablePadding: true,
-    label: '번호',
-  },
-  {
-    id: [MileageCategoryBoard.CATEGORY],
-    numeric: true,
-    disablePadding: false,
-    label: '카테고리명',
-  },
-  {
-    id: [MileageCategoryBoard.MAX_MILEAGE],
-    numeric: true,
-    disablePadding: false,
-    label: '최대 마일리지',
-  },
-  {
-    id: [MileageCategoryBoard.MANAGE],
-    numeric: true,
-    disablePadding: false,
-    label: '관리',
-  },
-  {
-    id: [MileageCategoryBoard.CHECK_BOX],
-    numeric: true,
-    disablePadding: false,
-    label: '체크',
-  },
-];
+// const headCells = [
+//   {
+//     id: [MileageCategoryBoard.NUM],
+//     numeric: false,
+//     disablePadding: true,
+//     label: '번호',
+//   },
+//   {
+//     id: [MileageCategoryBoard.CATEGORY],
+//     numeric: true,
+//     disablePadding: false,
+//     label: '카테고리명',
+//   },
+//   {
+//     id: [MileageCategoryBoard.MAX_MILEAGE],
+//     numeric: true,
+//     disablePadding: false,
+//     label: '최대 마일리지',
+//   },
+//   {
+//     id: [MileageCategoryBoard.MANAGE],
+//     numeric: true,
+//     disablePadding: false,
+//     label: '관리',
+//   },
+//   {
+//     id: [MileageCategoryBoard.CHECK_BOX],
+//     numeric: true,
+//     disablePadding: false,
+//     label: '체크',
+//   },
+// ];
 
 /**
  * @number 1번 목록
  * @description 마일리지 카테고리 리스트
  */
 
-const rows = [
-  createData(1, '전공 마일리지', 7, '웹 서비스 캠프', <StarIcon />),
-  createData(2, '비교과 - 연구활동', 6, '웹 서비스 캠프', <StarIcon />),
-  createData(3, '비교과 - 전공활동', 6, '웹 서비스 캠프', <StarIcon />),
-  createData(4, '비교과 - 특강참여', 7, '웹 서비스 캠프', <StarIcon />),
-  createData(5, '비교과 - 학회활동', 6, '웹 서비스 캠프', <StarIcon />),
-  createData(6, '비교과 - 행사참여', 8, '웹 서비스 캠프', <StarIcon />),
-];
+// const rows = [
+//   createData(1, '전공 마일리지', 7, '웹 서비스 캠프', <StarIcon />),
+//   createData(2, '비교과 - 연구활동', 6, '웹 서비스 캠프', <StarIcon />),
+//   createData(3, '비교과 - 전공활동', 6, '웹 서비스 캠프', <StarIcon />),
+//   createData(4, '비교과 - 특강참여', 7, '웹 서비스 캠프', <StarIcon />),
+//   createData(5, '비교과 - 학회활동', 6, '웹 서비스 캠프', <StarIcon />),
+//   createData(6, '비교과 - 행사참여', 8, '웹 서비스 캠프', <StarIcon />),
+// ];
 
 interface EnhancedTableProps {
   numSelected: number;
@@ -220,10 +220,12 @@ interface EnhancedTableProps {
   order: Order;
   orderBy: string;
   rowCount: number;
+  headCells: HeadCell[];
 }
 
 function EnhancedTableHead(props: EnhancedTableProps) {
-  const { onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort } = props;
+  const { headCells, onSelectAllClick, order, orderBy, numSelected, rowCount, onRequestSort } =
+    props;
   const createSortHandler = (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
     onRequestSort(event, property);
   };
@@ -317,7 +319,7 @@ function EnhancedTableToolbar(props: EnhancedTableToolbarProps) {
   );
 }
 
-export default function EnhancedTable({ type }: { type: string }) {
+export default function EnhancedTable({ type, rows, headCells }: { type: string; rows: any }) {
   const [order, setOrder] = React.useState<Order>('asc');
   const [orderBy, setOrderBy] = React.useState<keyof Data>('calories');
   const [selected, setSelected] = React.useState<readonly string[]>([]);
@@ -398,6 +400,7 @@ export default function EnhancedTable({ type }: { type: string }) {
             size={dense ? 'small' : 'medium'}
           >
             <EnhancedTableHead
+              headCells={headCells}
               numSelected={selected.length}
               order={order}
               orderBy={orderBy}
